Fixed regression with dashed bonds introduced in r863, see USRE039991-20080101-C00229.TIF
git-svn-id: svn://svn.code.sf.net/p/osra/code/trunk@869 049e15af-3938-0410-b28b-82a5db9f179f

+++ src/osra_structure.cpp
@@ -1239,8 +1239,8 @@ int find_dashed_bonds(const potrace_path_t *p, vector<atom_t> &atom, vector<bond
           n = p->curve.n;
           c = p->curve.c;
           int *tag = p->curve.tag;
-	  int cx =  c[n - 1][2].x;
-	  int cy =  c[n - 1][2].y;
+	  double cx =  c[n - 1][2].x;
+	  double cy =  c[n - 1][2].y;
 	  if (cx<0) cx=0;
 	  if (cx>width) cx=width;
 	  if (cy<0) cy=0;
